backend/requirements.txt
Django==4.1
djangorestframework==3.13.1
djoser==2.1.0
drf-extra-fields==3.4.0 
flake8==5.0.4
flake8-isort==4.2.0
flake8-return==1.1.3
pep8-naming==0.13.2
Pillow==9.2.0
pytest==7.1.2
